Use dataclasses for metadata
This is a purely back-end change; other than the new "not an entry" default of `MISSING` instead of `None`, this is not even visible to the user. No tests had to be adjusted; behaviour is preserved.

Buuuut, only because we manually include all the **kwargs used in the tests when annotating types -- this is not actually a robust and universal solution yet.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>