Test for type inheritance
At macro IO-subgraph interface. We still (correctly) get the connection flagged as invalid, but the exact description of _how_ it is invalid has changed. Instead of the downstream having an incommensurate type with the upstream, it has its original downstream type _and_ the upstream type simultaneously (which is not allowed).

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>